Fix fbt.subject
Summary:
I stumbled upon this when writing test for js fbt fbsource.

Instead of pushing subject call into runtimeArgs, we should do an unshift to prepend it to the array. This only fails when there are multiple _ calls which result in subject being the last instead of the first, but the table is still built as if subject is the first level.

Existing test cases are not covering this so added a test.

See updated snapshot for better explanation...

Reviewed By: jrwats

Differential Revision: D12973429

fbshipit-source-id: ac16034d7437d86f08a1974b0771da38642f10d6