Fix footer docs links on webpage (fixes #28) (#29)
Summary:
On footer all the links are using the docUrl method that is adding the language on the url:

/fbt/docs/en/api_intro
/fbt/docs/en/getting_started
However, these links are not correct. Getting an 404 error.

Although maybe in the future will be support for many languages, I think is better to change this invalid links for the valid links right now.
Pull Request resolved: https://github.com/facebookincubator/fbt/pull/29

Differential Revision: D13587298

fbshipit-source-id: 1cce44e8689c7ebc5e69643abae8029018fcd3c7