Use `shelljs` for Windows interoperability
Summary:
We can't run the `manfiest.js` util in Windows as is, since it relies on `find` and `grep` utilities.  So bring in the `shelljs` package to perform this.

We should consider revisiting this in the future should this solution be too slow.  As it is, this should be a performance hit compared to `find | grep` as we're essentially performing `fs.readFileSync(...)` over and over.

Reviewed By: w1106c

Differential Revision: D13625762

fbshipit-source-id: 30bec8dd80b7a1ed3e3353dfed719ee505eedb7f