Make translate.js script useable
Summary:
The `translate.js` script as it is, is not useable with collectFbt.js directly.  At the same, requiring people to aggregate all translations into one monolothic JSON file is cumbersome for most use-cases.

Instead, enable consuming these separately.  Note: this is a breaking change as we don't default to JSON coming from stdin anymore.  Instead the `--stdin` option is required to maintain backwards compatibility.

Included in this change is the `--pretty` option for both `collectFbt` and `translate` for pretty printing the JSON out.

Additionally, while I'm here, I'm getting `translate` off of the `optimist` command-line option/argument/command parser (it hasn't been in development for 5 years) and onto `yargs`, its successor.

Reviewed By: kayhadrin

Differential Revision: D13648187

fbshipit-source-id: d1b8d74eaebf13347f1a02349901505ede17e17b