Prepare fbt runtime for npm module
Summary: To publish to npm such that our fbt runtime is useable out-of-the-box on npm, we'll be making a "public-facing" module that will have fbt and any relevant modules that a client may want to reference (like our constants).  We are changing `FbtTranslations` to employ an initializer to register translations rather than relying on a magic static module `'src/translatedFbts.js'`.

Reviewed By: kayhadrin

Differential Revision: D13848879

fbshipit-source-id: f128d207d141c3b47513fee623ee82e98ec16b92