Prune fbt.plural branches with the same count arguments
Summary:
This prunes branches of duplicate `fbt.plural` entries (those using the same "count" argument) from `fbt` output.

Note that we *don't* prune runtime arguments of the fbt.plural calls and we leave the "levels" in the tree that *could* have been collapsed.  Why? I think we have some API-level decisions to make before we address that.  I'm tracking *that* work in T41100260

Reviewed By: kayhadrin, scobo

Differential Revision: D14256188

fbshipit-source-id: 0d32cd27d3cbce4bd6238f2095d7ad00d4d25110