Fix instructions in "Getting Started" (#57)
Summary:
Was trying out FBT and was on the "Getting Started" page and found that the instructions were not quite correct. Doing `yarn install` immediately after cloning isn't right as the user is not yet within the cloned repo. I updated the steps and managed to get the demo app running.
Pull Request resolved: https://github.com/facebookincubator/fbt/pull/57

Reviewed By: yangshun

Differential Revision: D14740746

Pulled By: yangshun

fbshipit-source-id: bf84888e2946b62e5f9a9b3e602c520fe892557e