Enable custom Babel plugins for FBT collection
Summary:
:
FBT collection relies on manually running the Babel parser.  Ideally, we'd just pull from `.babelrc` or the same plugin list as the calling `babel-node` up the stack, and clients could just keep their "source of truth" for Babel plugins in one spot.  This is a stop-gap for now to allow callers to pass in Babel plugins they want to enable.

Reviewed By: kayhadrin

Differential Revision: D17071522

fbshipit-source-id: b39e512faab7a0172b5f22ef5c33860084ce6aa2