New arg split-by-locale-to added to translate.js (#73)
Summary:
This PR adds the functionality to `translate.js` to split the translations into separate files based on local. The arg to do so is called 'split-by-locale-to', and I'm open for suggestions on the naming of the arg. This functionality will reduce the manual work of code splitting / lazy load translations based on locale.

The PR was based on the discussions in [this](https://github.com/facebookincubator/fbt/issues/71) issue. However, the implementation differs a bit from the solution presented [here](https://github.com/facebookincubator/fbt/issues/71#issuecomment-525994553). The reason for the difference is that `translate.js` outputs default to stdout, which would be messy to deal with when we want to output to several files. The solution [here](https://github.com/facebookincubator/fbt/issues/71#issuecomment-525994553) could have been used if a new script was created, however in my opinion it seems like a better option to add an arg instead to the existing script.

NOTE: I'm probably overseeing something, but I was not able to run any linter or any tests for the project before the commit. Seems like there are no eslint config? And there are only test scripts and a jest config added for the demo app?
Pull Request resolved: https://github.com/facebookincubator/fbt/pull/73

Reviewed By: jrwats

Differential Revision: D17222829

Pulled By: kayhadrin

fbshipit-source-id: d180597f9eb5003ce5402634c8d98cd59b5a3fc2