Port Tiger hash to native JS
Summary:
HHVM support in fbsource is going away, and we soon won't be able to run `tiger128.php` outside of `www`.

The internal i18n hashing relies on the bespoke `tiger-3,fb` hash, which replicates the PHP's old tiger implementation that had a swapped byte order.  Port the implementation over to native JavaScript using `BigInt` (available in Node v10.4.0 and above) so that it can be used without shelling out to HHVM's implementation

Adding this as its own NPM package so we don't bloat `babel-plugin-fbt` too much and so that using the CLI FBT collect-fbts command is as straightforward as the following (when installed via npm):
```
node node_modules/babel-plugin-fbt/bin/collectFBT --hash-module 'fb-tiger/hashPhrases' ...
```

Reviewed By: kayhadrin

Differential Revision: D17631968

fbshipit-source-id: 8e250cc44217cfe35dbad2b86a6af1b73ad68b14