chore: [EZ] Align FbtHooks with implicit-inexact-object
Summary:
`Shape<T>` appears to be a better type for "option bags" (where there are no required keys).  It both:
 * Won't let you assign to keys that aren't in its set (like an exact type)
 * *Will* allow object literals (for initializing)

When syncing to fbsource, this type was causing issues as the empty object literal is inexact, and `FbtHookRegistrations` is exact.

If we ever have a required key, we can go back to objects.

```
export type FbtHookRegistrations = {
  errorListener: (context: FbtErrorContext) => IFbtErrorListener,
  logImpression: (hash: string) => void,
  overrideListener: (context: FbtErrorContext) => IFbtErrorListener,
  onOverride: (hash: string) => void,
  ...requiredKeys...,
};
const _registrations: FbtHookRegistrations = {...requiredKeys...};
```

Reviewed By: kayhadrin

Differential Revision: D19085299

fbshipit-source-id: afbcbe2b341c331641012d1c15cf7d44403739f9