feat(Fbt): Expose FbtResult in NPM package
Summary:
It's completely within reason that someone may want to override `getFbtResult` in `FbtHooks`, but also defer to the provided `FbtResult` in certain cases.  Without exposing the class, there's no way for someone do so.

Additionally, `FbtResult` comes with its own set of helpers, such as the flatten functions.  So let's expose it.

Reviewed By: dalmendray

Differential Revision: D19602003

fbshipit-source-id: f6232d2866f5c2035423b3c512ad2e82b8cb1adf