docs(website): add DocSearch as recommended by docusaurus. (#130)
Summary:
👋 team,

I am working on DocSearch. [We have an integration for Docusaurus](https://docusaurus.io/docs/en/search#enabling-the-search-bar). We thought it is a shame you can not also used this search that you can [find on reactjs for example](https://reactjs.org/).

This PR will add DocSearch to the documentation website. It will allow an user to have a learn-as-you-type experience by displaying results thanks to a dropdown in a live way.

Let me know if you need anything.
Pull Request resolved: https://github.com/facebookincubator/fbt/pull/130

Differential Revision: D20040157

Pulled By: jrwats

fbshipit-source-id: e442b1f4fa40e5e0cbd96e1d3fb018b4d587178c