fix(fbt): Rewrite our required modules in `*.js.flow`
Summary:
We use gulp to process our runtime source code for publishing the `fbt` package.  We perform the following on our source intended for runtime:
* flatten the directory structure,
* Strip flow types & **rewrite module names** (most relevant to this diff)
* create a minified `dist/` package,

For flow definitions, we just copy files (renamed with the `.flow` suffix) to the `/lib` dir, BUT the modules in the `import` and `require` statements are still accessed as "global" Haste module paths (unlike the flow-stripped source files, whose dependencies we *do* rewrite). If anyone attempts to use these `*.flow` files they break or result in worthless `any` types.  Let's fix that, so clients who want to use our types can do so.

Employ a flow-aware gulp-plugin that uses the `flow-parser` to scan the AST for `require` and `import` statements and replace  modules in in a manner similar to `babel-preset-fbjs/plugins/rewrite-modules` that we use for the flow-stripped source

Reviewed By: kayhadrin, scobo

Differential Revision: D20243589

fbshipit-source-id: e608b7b6471e13940fa6dac1cdb1866e0be03d4e