added __self to fbt ignored props (#137)
Summary:
Recently, a bugfix went out for `babel-plugin-transform-react-jsx-self` that moves adding the `__self` JSX attribute before the arrow function transform. This has the side effect of adding the `__self` attribute before the fbt transform as well. This PR marks `__self` as a passed through but purposefully ignored attribute so that this change does not cause fbt to throw because of an unknown attribute
Pull Request resolved: https://github.com/facebook/fbt/pull/137

Test Plan: I wrote a test to verify that the attribute is ignored. I also manually changed the code in the WWW repo and verified running the `self` transform in conjunction with this transform no longer throws an error.

Reviewed By: kayhadrin

Differential Revision: D20932477

Pulled By: lunaruan

fbshipit-source-id: 9d29608fc392ef962749d292a74d0c660d72b310