Upgrade webpack. Remove uglify
Summary:
I'm mostly doing this to get our security vulnerabilities green again.  This has a very minor affect on our built fbt distribution.  It's not really *that* impactful, since this "vulnerability" only affects the "build" of our library distribution, but still...

`uglifyjs-webpack-plugin` brings in `serialize-javascript@1.7.0`, which is woefully out of date and a security vulnerability.  It also adds hardly any value over the built-in `terser` plugin that comes with `webpack 5.0+`.  So let's upgrade webpack, and drop `uglify`.

Reviewed By: dalmendray

Differential Revision: D24862524

fbshipit-source-id: 063a2a723b3c093ef81373bec57dc99c410a0cce