fix(fbt): amend Flow typing of intlList()
Summary:
Previously, the exported Flow type of `intlList` was an intersection between a function and an object type:
```
module.exports = (intlList: (<TItem: React.Node>(
  items: $ReadOnlyArray<TItem>,
  conjunction: ?$Keys<typeof CONJUNCTIONS>,
  delimiter: ?$Keys<typeof DELIMITERS>,
) => TItem | Fbt) & {
  DELIMITERS: typeof DELIMITERS,
  CONJUNCTIONS: typeof CONJUNCTIONS,
  ...
});
```

I'm not entirely sure about the specifics around how Flow handles this, but it seems like Flow is just treating it as `nothing`, as it appears as `nothing` when I hover over it in VS Code.

This means that all call sites to `intlList` essentially had no Flow coverage at all. :(

I've fixed the type so that it accurately reflects the actual exports of the module. This resulted in several Flow errors across the codebase. For places where it was possible to fix these errors **with no runtime changes** (ie. Flow types only), I adjusted the types. The most common issue was call sites expecting this function to return a plain `string`, but in reality it returns an `Fbt` string.

There are several call sites that can't be fixed without changing the runtime behaviour, so I just left those as-is and added `$FlowFixMe` comments. The most common issues were:

 - Passing `null`s to the function even though it's typed to not accept nulls.
 - Doing string concatenation with the returned value, for example:
```
`Stuff: ${intlList(items)}`
'Stuff: ' + intlList(items)
```
This is not valid for `Fbt`'d strings. Concatenation should be done using an `<fbt>` string with `<fbt:param>`s, as different languages may have different sentence orders

I'd like to revamp this more (for example, use ES6 exports, and Flow enums), however this code is exported to open-source so I need to be careful about any changes.

I added `$FlowFixMe` comments using the [Flow tool](https://medium.com/flow-type/upgrading-flow-codebases-40ef8dd3ccd8):
```
./scripts/flow/tool add-comments . --comment "Found when typing intlList" --all
```

Reviewed By: kayhadrin

Differential Revision: D25698393

fbshipit-source-id: 74ad206fb96268f13a778a45834fdc6f67222968