feat(babel-plugin-fbt): inline fbjs dependency (#330)
Summary:
<!--
Thanks for submitting a pull request! Please provide enough
information so that others can review your pull request. The two
fields below are mandatory.
-->

<!--
Please remember to update CHANGELOG.md in the root of the project if
you have not done so.
-->

Closes https://github.com/facebook/fbt/issues/321

I had to move the comments in the `.gitignore` to a separate line, otherwise the dir wouldn't be ignored for me. Not sure if there is anything on my side that is broken?

I inlined `keyMirror` but `forEachObject` I replaced with `Object.keys` and a `for of` loop

Pull Request resolved: https://github.com/facebook/fbt/pull/330

Test Plan:
```
yarn clean-test
```

Reviewed By: kayhadrin

Differential Revision: D33735312

Pulled By: pkqinys

fbshipit-source-id: 26b52da543fa91f5069eb9576b5c4f0a1877e1d2