docs: add social banner in support of Ukraine (#347)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156671819-63e20333-6d64-452a-93f5-e037d8964ab2.png)

Pull Request resolved: https://github.com/facebook/fbt/pull/347

Reviewed By: pkqinys

Differential Revision: D34628776

Pulled By: dmitryvinn-fb

fbshipit-source-id: 71a4783286c73ec62be5919eb5eba3e168f79aea