Expose fbs API to OSS
Summary:
The `fbs` API was only partially exposed to the OSS version of the library; but accessing it required a separate import/require statement.

Let's bundle the `fbs.js` lib with the regular `fbt.js` since it's fairly lightweight.

I've also updated the demo app example to show how it's used.

Differential Revision: D35457016

fbshipit-source-id: 59d202c21ce0c55fa05af977902a4e8fba427f3c